Add Family to CashAccount. Store current family in session.

+++ HomERP.Domain/Logic/CashAccountProvider.cs
@@ -4,27 +4,52 @@ using System.Text;
 using HomERP.Domain.Entity;
 using HomERP.Domain.Logic.Abstract;
 using HomERP.Domain.Repository.Abstract;
+using System.Linq;
+using HomERP.Domain.Authentication;
 
 namespace HomERP.Domain.Logic
 {
     public class CashAccountProvider : ICashAccountProvider
     {
         private ICashAccountRepository repository;
-        public CashAccountProvider(ICashAccountRepository repository)
+        private ISessionDataProvider sessionProvider;
+        private Family family;
+        public CashAccountProvider(ICashAccountRepository repository, ISessionDataProvider sessionProvider)
         {
             this.repository = repository;
+            this.sessionProvider = sessionProvider;
+            this.Family = sessionProvider.GetFamily();
         }
+
+        public Family Family
+        {
+            set
+            {
+                if (value.Id <= 0) throw new ArgumentException();
+                this.family = value;
+            }
+        }
+
         public IEnumerable<CashAccount> CashAccounts
-        { get { return repository.CashAccounts; } }
+        { get { return repository.CashAccounts.Where(ca => ca.Family.Id == family.Id); } }
 
         public CashAccount DeleteCashAccount(int cashAccountId)
         {
-            return repository.DeleteCashAccount(cashAccountId);
+            var cashAccountToDelete = repository.CashAccounts.Where(ca => ca.Id == cashAccountId && ca.Family.Id == this.family.Id).FirstOrDefault();
+            if (cashAccountToDelete.Id == cashAccountId)
+            {
+                return repository.DeleteCashAccount(cashAccountId);
+            }
+            return null;
         }
 
         public void SaveCashAccount(CashAccount cashAccount)
         {
-            repository.SaveCashAccount(cashAccount);
+            if (cashAccount.Id == 0) cashAccount.Family = this.family;
+            if (cashAccount.Family.Id == this.family.Id)
+            {
+                repository.SaveCashAccount(cashAccount);
+            }
         }
     }
 }
